Struct bloguen::ops::BloguePost

pub struct BloguePost {
    pub source_dir: (String, PathBuf),
    pub number: (usize, String),
    pub name: String,
    pub datetime: DateTime<LocalOffset>,
}

Information about a blogue post.

Use list() to find valid post directories, then use new() to get the post data.

A correct post directory name is #+. YYYY-MM-DD [HH-MM[-SS]] name.

Fields
source_dir: (String, PathBuf)

Directory containing the post data.

number: (usize, String)

Post number.

name: String

Post name.

datetime: DateTime<LocalOffset>

Date & time of posting.

Methods

impl BloguePost

pub fn list(within: &(String, PathBuf)) -> Result<Vec<(String, PathBuf)>, Error>

List correct post directories in the specified directory.

A correct post directory name is #+. YYYY-MM-DD [HH-MM[-SS]] name.

Examples:

For the following directory tree:

posts/
  temp/
  001. 2018-01-08 16-52 The venture into crocheting/
  002. 2018-01-08 acquiescence.md
  003. 2018-02-05 release-front - release front-end/
  004. stir plate/
  blogue.toml

The following holds:

let root: PathBuf = /* obtained elsewhere */;
let posts = BloguePost::list(&("$ROOT/posts/".to_string(), root.join("posts"))).unwrap();
assert_eq!(&posts[..],
    &[("$ROOT/posts/001. 2018-01-08 16-52 The venture into crocheting/".to_string(),
       root.join("posts").join("001. 2018-01-08 16-52 The venture into crocheting")),
      ("$ROOT/posts/003. 2018-02-05 release-front - release front-end/".to_string(),
       root.join("posts").join("003. 2018-02-05 release-front - release front-end"))][..]);

pub fn new(wher: (String, PathBuf)) -> Result<BloguePost, Error>

Read post data into a BloguePost instance.

Examples:

let root: PathBuf = /* obtained elsewhere */;

let dir = ("$ROOT/posts/01. 2018-01-08 16-52 The venture into crocheting".to_string(),
           root.join("posts").join("01. 2018-01-08 16-52 The venture into crocheting"));
assert_eq!(BloguePost::new(dir.clone()),
           Ok(BloguePost {
               source_dir: dir,
               number: (1, "01".to_string()),
               name: "The venture into crocheting".to_string(),
               datetime: LocalOffset.ymd(2018, 01, 08).and_hms(16, 52, 00),
           }));

let dir = ("$ROOT/posts/003. 2018-02-05 release-front - release front-end".to_string(),
           root.join("posts").join("003. 2018-02-05 release-front - release front-end"));
assert_eq!(BloguePost::new(dir.clone()),
           Ok(BloguePost {
               source_dir: dir,
               number: (3, "003".to_string()),
               name: "release-front - release front-end".to_string(),
               datetime: LocalOffset.ymd(2018, 02, 05).and_hms(23, 24, 43),
           }));

let dir = ("$ROOT/posts/004. stir plate".to_string(),
           root.join("posts").join("004. stir plate"));
assert_eq!(BloguePost::new(dir.clone()),
           Err(Error::Parse {
               tp: "post directory filename",
               wher: "blogue post".into(),
               more: None,
           }));

pub fn generate(
&self,
into: &(String, PathBuf),
post_header: &str,
post_footer: &str,
blog_name: &str,
language: &LanguageTag,
author: &str,
spec_tags: &[TagName],
free_tags: &[TagName],
post_data: &BTreeMap<String, String>,
global_data: &BTreeMap<String, String>,
post_styles: &[StyleElement],
global_styles: &[StyleElement],
post_scripts: &[ScriptElement],
global_scripts: &[ScriptElement]
) -> Result<Vec<String>, Error>

Generate an HTML output from the post into the specified output directory.

Returns: set of links in the markdown source.

Examples

Given the following:

src/
  01. 2018-01-08 16-52 The venture into crocheting/
    post.md

The following holds:

let root: PathBuf = /* obtained elsewhere */;
let post =
    BloguePost::new(("$ROOT/src/01. 2018-01-08 16-52 The venture into crocheting".to_string(),
        root.join("src").join("01. 2018-01-08 16-52 The venture into crocheting"))).unwrap();
assert!(post.generate(&("$ROOT/out/".to_string(), root.join("out")), "header", "footer",
                      "Блогг", &LANGUAGE_EN_GB, "autheur", &[], &[], &Default::default(), &Default::default(),
                      &[], &[], &[], &[]).is_ok());

assert!(root.join("out").join("posts")
            .join("01. 2018-01-08 16-52-00 The venture into crocheting.html").is_file());

pub fn generate_machine(
&self,
into: &(String, PathBuf),
subpath: &str,
kind: &MachineDataKind,
blog_name: &str,
language: &LanguageTag,
author: &str,
spec_tags: &[TagName],
free_tags: &[TagName],
post_data: &BTreeMap<String, String>,
global_data: &BTreeMap<String, String>,
post_styles: &[StyleElement],
global_styles: &[StyleElement],
post_scripts: &[ScriptElement],
global_scripts: &[ScriptElement]
) -> Result<(), Error>

Generate machine output of the specified kind from the post into the specified subpath in the specified output directory.

Examples

Given the following:

src/
  01. 2018-01-08 16-52 The venture into crocheting/
    post.md

The following holds:

let root: PathBuf = /* obtained elsewhere */;
let post =
    BloguePost::new(("$ROOT/src/01. 2018-01-08 16-52 The venture into crocheting".to_string(),
        root.join("src").join("01. 2018-01-08 16-52 The venture into crocheting"))).unwrap();
assert!(post.generate_machine(&("$ROOT/out/".to_string(), root.join("out")), "machine/", &MachineDataKind::Json,
                              "Блогг", &LANGUAGE_EN_GB, "autheur", &[], &[], &Default::default(), &Default::default(),
                      &[], &[], &[], &[]).is_ok());

assert!(root.join("out").join("machine")
            .join("01. 2018-01-08 16-52-00 The venture into crocheting.json").is_file());

pub fn normalised_name(&self) -> String

Get a normalised output name for this post.

Examples

let root: PathBuf = /* obtained elsewhere */;
let post =
    BloguePost::new(("$ROOT/src/01. 2018-01-08 16-52 The venture into crocheting".to_string(),
        root.join("src").join("01. 2018-01-08 16-52 The venture into crocheting"))).unwrap();
assert_eq!(post.normalised_name(), "01. 2018-01-08 16-52-00 The venture into crocheting");

pub fn copy_asset(
&self,
into: &(String, PathBuf),
link: &str
) -> Result<bool, Error>

Copy a referenced asset to the output directory.

Returns Ok(b), where b is whether the asset existed and was copied, Err(_) for a copying error.

Examples

let root: PathBuf = /* obtained elsewhere */;
let out_pair = ("$ROOT/out/".to_string(), root.join("out"));
let post =
    BloguePost::new(("$ROOT/src/01. 2018-01-08 16-52 The venture into crocheting".to_string(),
        root.join("src").join("01. 2018-01-08 16-52 The venture into crocheting"))).unwrap();
for link in post.generate(&out_pair, "header", "footer", "Блогг", &LANGUAGE_EN_GB, "autheur",
                          &[], &[], &Default::default(), &Default::default(), &[], &[], &[], &[])
            .unwrap().into_iter().filter(|l| util::is_asset_link(l)) {
    let link = percent_decode(link.as_bytes()).decode_utf8().unwrap();
    println!("Copying {}: {:?}", link, post.copy_asset(&out_pair, &link));
}